+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "fe-data-fetching",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "app.js",
   "scripts": {
     "start": "nodemon"
   },
